requirements.txt
paddlepaddle
paddleocr
pillow
setuptools
fastapi
uvicorn
python-multipart